feat: update deployer with cmd
