+++ .env.example
@@ -3,15 +3,27 @@ STARKNET_ADMIN_ADDRESS_DEV=0x517ececd29116499f4a1b64b094da79ba08dfd54a3edaa31613
 STARKNET_ADMIN_PRIVATE_KEY_DEV=0x1800000000300000180000000000030000000000003006001800006600
 STARKNET_ADMIN_PUBLIC_KEY_DEV=0x2b191c2f3ecf685a91af7cf72a43e7b90e2e41220175de5c4f7498981b10053
 
+STARKNET_SOLIS_ACCOUNT_ADDRESS_DEV=0x517ececd29116499f4a1b64b094da79ba08dfd54a3edaa316134c41f8160973
+STARKNET_SOLIS_ACCOUNT_PRIVATE_KEY_DEV=0x1800000000300000180000000000030000000000003006001800006600
+STARKNET_SOLIS_ACCOUNT_PUBLIC_KEY_DEV=0x2b191c2f3ecf685a91af7cf72a43e7b90e2e41220175de5c4f7498981b10053
+
 # Starknet account goerli
 STARKNET_ADMIN_ADDRESS_GOERLI=0x517ececd29116499f4a1b64b094da79ba08dfd54a3edaa316134c41f8160973
 STARKNET_ADMIN_PRIVATE_KEY_GOERLI=0x1800000000300000180000000000030000000000003006001800006600
 STARKNET_ADMIN_PUBLIC_KEY_GOERLI=0x2b191c2f3ecf685a91af7cf72a43e7b90e2e41220175de5c4f7498981b10053
 
+STARKNET_SOLIS_ACCOUNT_ADDRESS_GOERLI=
+STARKNET_SOLIS_ACCOUNT_PRIVATE_KEY_GOERLI=
+STARKNET_SOLIS_ACCOUNT_PUBLIC_KEY_GOERLI=
+
 # Starknet account mainnet
 STARKNET_ADMIN_ADDRESS_MAINNET=0x517ececd29116499f4a1b64b094da79ba08dfd54a3edaa316134c41f8160973
 STARKNET_ADMIN_PRIVATE_KEY_MAINNET=0x1800000000300000180000000000030000000000003006001800006600
 STARKNET_ADMIN_PUBLIC_KEY_MAINNET=0x2b191c2f3ecf685a91af7cf72a43e7b90e2e41220175de5c4f7498981b10053
+
+STARKNET_SOLIS_ACCOUNT_ADDRESS_MAINNET=
+STARKNET_SOLIS_ACCOUNT_PRIVATE_KEY_MAINNET=
+STARKNET_SOLIS_ACCOUNT_PUBLIC_KEY_MAINNET=
 
 # Solis accounts dev
 SOLIS_ADMIN_ADDRESS_DEV=0x517ececd29116499f4a1b64b094da79ba08dfd54a3edaa316134c41f8160973
@@ -28,7 +40,12 @@ SOLIS_ADMIN_ADDRESS_MAINNET=0x517ececd29116499f4a1b64b094da79ba08dfd54a3edaa3161
 SOLIS_ADMIN_PRIVATE_KEY_MAINNET=0x1800000000300000180000000000030000000000003006001800006600
 SOLIS_ADMIN_PUBLIC_KEY_MAINNET=0x2b191c2f3ecf685a91af7cf72a43e7b90e2e41220175de5c4f7498981b10053
 
-# Starknet network
-STARKNET_NETWORK_ID=dev
-# Solis network
-SOLIS_NETWORK_ID=dev
+# Starknet RPC
+STARKNET_NODE_URL_DEV=http://0.0.0.0:5050
+STARKNET_NODE_URL_GOERLI=https://starknet-testnet.public.blastapi.io
+STARKNET_NODE_URL_MAINNET=https://starknet-mainnet.public.blastapi.io
+
+# Solis RPC
+SOLIS_NODE_URL_DEV=http://0.0.0.0:7777
+SOLIS_NODE_URL_GOERLI=staging.solis.arkproject.dev
+SOLIS_NODE_URL_MAINNET=https://solis.arkproject.dev
