+++ contracts.json
@@ -1,1 +1,1 @@
-{"goerli":{"nftContract":"0x78e76f951aee17ab6f2075987890e0493cf7a8d14243c278729b1cead193c20","messaging":"0x25eab997d5bf56abdb3ce33d873997525eea99fa9c527f0809df09bbbcc890","executor":"0x68c317e638e8446dee9638089e4de00c1e7750174016796bf12d316c5d4942c"},"sepolia":{},"mainnet":{},"dev":{"orderbook":"0xbe5bf705ebf2b4219b4acd68c65daf3f13d41006de4868f27dbe98ae6206a7"}}
+{"goerli":{},"sepolia":{},"mainnet":{},"dev":{"messaging":"0x4c1a03668c2400873320a6adb7dc8f909792dea18a3e62ae683dc0cc227106d","executor":"0x224bfd2c4de87a8ae48ffbdf4b9ab32e98b5c21bcb3bb04ad0e4c53676d67a3","nftContract":"0x2963af055c65fa5c62fb9e8928fa8b53c8c71815ca30334628fc21a03a48d50","eth":"0x172ca09df59789985a1dfb68e0f3f554121eb313c68cf391bd78f9bf69abcf","orderbook":"0x3b2935da28f3c79abb8aadb39b27c012e6c6aef7e196a7de5671a8664dfec0e"}}
