+++ crates/solis/messaging.local.json
@@ -1,9 +1,9 @@
 {
   "chain": "starknet",
   "rpc_url": "http://127.0.0.1:5050",
-  "contract_address": "0x59b737fc89b970cb16260f36f08048267aeb9f3e8008c45c9e4a0f0deb54e3d",
-  "sender_address": "0x517ececd29116499f4a1b64b094da79ba08dfd54a3edaa316134c41f8160973",
-  "private_key": "0x1800000000300000180000000000030000000000003006001800006600",
+  "contract_address": "0x4c1a03668c2400873320a6adb7dc8f909792dea18a3e62ae683dc0cc227106d",
+  "sender_address": "0x5686a647a9cdd63ade617e0baf3b364856b813b508f03903eb58a7e622d5855",
+  "private_key": "0x33003003001800009900180300d206308b0070db00121318d17b5e6262150b",
   "interval": 2,
   "from_block": 0
 }
