+++ packages/deployer/package.json
@@ -22,7 +22,7 @@
     "dotenv": "^16.3.1",
     "loading-cli": "^1.1.2",
     "node-fetch": "^3.3.2",
-    "starknet": "5.24.3"
+    "starknet": "5.22.0"
   },
   "devDependencies": {
     "@types/node": "^20.10.1",
