+++ packages/deployer/src/deployStarknetFreemintNft.ts
@@ -1,36 +1,34 @@
 import { promises as fs } from "fs";
 
+import { program } from "commander";
+import loading from "loading-cli";
 import { CallData, Contract } from "starknet";
 
-import { SOLIS_NETWORK, STARKNET_NETWORK } from "./constants";
+import { ARTIFACTS_PATH } from "./constants";
 import { loadArtifacts } from "./contracts/common";
 import { deployERC20 } from "./contracts/erc20";
-import { getProvider } from "./providers";
+import { getStarknetProvider } from "./providers";
+import { ProviderNetwork } from "./types";
 import {
   getContractsFilePath,
-  getExistingAccounts,
-  getExistingContracts
+  getExistingContracts,
+  getStarknetAccounts
 } from "./utils";
 
-const loading = require("loading-cli");
-const artifactsPath = "../../contracts/target/dev/";
-
-export async function deployStarknetContracts() {
-  const { starknetProvider } = getProvider(STARKNET_NETWORK, SOLIS_NETWORK);
-  const { starknetAdminAccount } = getExistingAccounts(
-    STARKNET_NETWORK,
-    SOLIS_NETWORK
-  );
-
+export async function deployStarknetContracts(
+  starknetNetwork: ProviderNetwork
+) {
+  const starknetProvider = getStarknetProvider(starknetNetwork);
+  const { starknetAdminAccount } = getStarknetAccounts(starknetNetwork);
   let existingContracts = await getExistingContracts();
 
   console.log("\nSTARKNET ACCOUNTS");
   console.log("=================\n");
-  console.log(`| Admin account |  ${starknetAdminAccount.account.address}`);
+  console.log(`| Admin account |  ${starknetAdminAccount.address}`);
   console.log("");
 
   const starknetSpinner = loading("Deploying Nft Contract...").start();
-  const artifacts = loadArtifacts(artifactsPath, "ark_tokens_FreeMintNFT");
+  const artifacts = loadArtifacts(ARTIFACTS_PATH, "ark_tokens_FreeMintNFT");
 
   const contractCallData = new CallData(artifacts.sierra.abi);
   const contractConstructor = contractCallData.compile("constructor", {
@@ -38,7 +36,7 @@ export async function deployStarknetContracts() {
     symbol: "ARK"
   });
 
-  const deployR = await starknetAdminAccount.account.declareAndDeploy({
+  const deployR = await starknetAdminAccount.declareAndDeploy({
     contract: artifacts.sierra,
     casm: artifacts.casm,
     constructorCalldata: contractConstructor
@@ -52,8 +50,8 @@ export async function deployStarknetContracts() {
 
   existingContracts = {
     ...existingContracts,
-    [STARKNET_NETWORK]: {
-      ...existingContracts[STARKNET_NETWORK],
+    [starknetNetwork]: {
+      ...existingContracts[starknetNetwork],
       nftContract: nftContract.address
     }
   };
@@ -61,12 +59,12 @@ export async function deployStarknetContracts() {
   await fs.writeFile(getContractsFilePath(), JSON.stringify(existingContracts));
 
   let ethContract: Contract | undefined;
-  if (STARKNET_NETWORK === "dev") {
+  if (starknetNetwork === "dev") {
     starknetSpinner.text = "Deploying Eth Contract...";
 
     ethContract = await deployERC20(
-      artifactsPath,
-      starknetAdminAccount.account,
+      ARTIFACTS_PATH,
+      starknetAdminAccount,
       starknetProvider,
       "ETH",
       "ETH"
@@ -74,8 +72,8 @@ export async function deployStarknetContracts() {
 
     existingContracts = {
       ...existingContracts,
-      [STARKNET_NETWORK]: {
-        ...existingContracts[STARKNET_NETWORK],
+      [starknetNetwork]: {
+        ...existingContracts[starknetNetwork],
         eth: ethContract.address
       }
     };
@@ -93,4 +91,10 @@ export async function deployStarknetContracts() {
   }
 }
 
-deployStarknetContracts();
+program.option("-sn, --starknet <type>", "Starknet Network", "dev");
+program.parse();
+
+const options = program.opts();
+const starknetNetwork = options.starknet;
+
+deployStarknetContracts(starknetNetwork);
