+++ packages/deployer/src/providers.ts
@@ -1,22 +1,7 @@
 import { RpcProvider } from "starknet";
 
-export function getProvider(starknetNetwork: string, solisNetwork: string) {
-  let solisNodeUrl: string;
+export function getStarknetProvider(starknetNetwork: string) {
   let starknetNodeUrl: string;
-  switch (solisNetwork) {
-    case "dev":
-      solisNodeUrl = "http://0.0.0.0:7777";
-      break;
-    case "mainnet":
-      solisNodeUrl = "solis.arkproject.dev";
-      break;
-    case "goerli":
-      solisNodeUrl = "staging.solis.arkproject.dev";
-      break;
-    default:
-      throw new Error(`Unsupported solisNetwork: ${solisNetwork}`);
-  }
-
   switch (starknetNetwork) {
     case "dev":
       starknetNodeUrl = process.env.STARKNET_NODE_URL_DEV || "";
@@ -31,16 +16,31 @@ export function getProvider(starknetNetwork: string, solisNetwork: string) {
       throw new Error(`Unsupported starknetNetwork: ${starknetNetwork}`);
   }
 
-  const solisProvider = new RpcProvider({ nodeUrl: solisNodeUrl });
-  const starknetProvider = new RpcProvider({ nodeUrl: starknetNodeUrl });
+  return new RpcProvider({ nodeUrl: starknetNodeUrl });
+}
 
-  return { solisProvider, starknetProvider };
+export function getSolisProvider(solisNetwork?: string) {
+  let solisNodeUrl: string;
+  switch (solisNetwork) {
+    case "dev":
+      solisNodeUrl = process.env.SOLIS_NODE_URL_DEV || "";
+      break;
+    case "mainnet":
+      solisNodeUrl = process.env.SOLIS_NODE_URL_GOERLI || "";
+      break;
+    case "goerli":
+      solisNodeUrl = process.env.SOLIS_NODE_URL_MAINNET || "";
+      break;
+    default:
+      throw new Error(`Unsupported solisNetwork: ${solisNetwork}`);
+  }
+
+  return new RpcProvider({ nodeUrl: solisNodeUrl });
 }
 
 export function getFeeAddress(network: string) {
   switch (network) {
-    case "katana":
-      // TODO use deployed ERC20 contract address
+    case "dev":
       return "0x49d36570d4e46f48e99674bd3fcc84644ddd6b96f7c741b1562b82f9e004dc7";
     case "goerli":
       return "0x049d36570d4e46f48e99674bd3fcc84644ddd6b96f7c741b1562b82f9e004dc7";
